refactor: 각 예외들 구분해주는 폴더 생성
