+++ src/main/java/kuit/subway/service/StationService.java
@@ -1,16 +1,13 @@
 package kuit.subway.service;
 
-import jakarta.persistence.EntityNotFoundException;
 import kuit.subway.domain.Station;
 import kuit.subway.dto.request.station.StationCreateRequest;
 import kuit.subway.dto.response.station.StationCreateResponse;
 import kuit.subway.dto.response.station.StationDeleteResponse;
 import kuit.subway.dto.response.station.StationDto;
-import kuit.subway.exception.notfound.NotFoundException;
-import kuit.subway.exception.notfound.NotFoundStationException;
+import kuit.subway.exception.notfound.station.NotFoundStationException;
 import kuit.subway.repository.StationRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.cglib.core.Local;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
